third_party/WebKit/LayoutTests/accessibility/aria-option-role-expected.txt

option 1
option 2

This tests that the aria 'option' role works as expected. That is, it becomes a static text element with no children.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

firstChild.role is AXRole: AXListBoxOption
PASS firstChild.name is 'option 1'
secondChild.role is AXRole: AXListBoxOption
PASS secondChild.name is 'label 2'
PASS firstChild.childrenCount is 0
PASS secondChild.childrenCount is 0
PASS firstChild.isFocusable is true
PASS successfullyParsed is true

TEST COMPLETE
